Main Javascript Functionality Done
Setup for potentially all php files are ready to go. Most of these are gonna be setting up a database or acting as a helper function to what the javascript needs. Figured it was best to hit the main critical javascript functionality like messages and whatnot but I'll review what could be missing here next.